Remove unnecessary uses of `useCallback`
We copied and pasted a lot of code that used `useCallback` into places where
this wasn't needed to prevent re-renders.

- Remove `useCallback` from lots of functions where the performance
  optimization isn't necessary
- Refactor `useFileNameAndExtension` hook into `parseFileNameAndExtension`
  function
