+++ src/components/Formatters/index.ts
@@ -1,3 +1,4 @@
 export * from './useDateFormatter';
 export * from './useNumberFormatter';
 export * from './formatFileSize';
+export * from './parseFileNameAndExtension';
