+++ src/components/files/FilePreview/useRetrieveFile.ts
@@ -1,40 +1,36 @@
 import { useSnackbar } from 'notistack';
-import { useCallback } from 'react';
 import { SupportedType } from '../FILE_MIME_TYPES';
 
 export const useRetrieveFile = () => {
   const { enqueueSnackbar } = useSnackbar();
 
-  const showSnackbarError = useCallback(() => {
+  const showSnackbarError = () => {
     enqueueSnackbar('Could not retrieve file', {
       variant: 'error',
     });
-  }, [enqueueSnackbar]);
+  };
 
-  const retrieveFile = useCallback(
-    async (
-      url: string,
-      type: SupportedType,
-      callback: (file: File) => void | Promise<void>,
-      errorHandler?: () => void
-    ): Promise<void> => {
-      try {
-        const response = await fetch(url);
-        if (response.status === 200) {
-          void callback(
-            new File([await response.blob()], 'Preview', {
-              type,
-            })
-          );
-        } else {
-          errorHandler ? errorHandler() : showSnackbarError();
-        }
-      } catch {
+  const retrieveFile = async (
+    url: string,
+    type: SupportedType,
+    callback: (file: File) => void | Promise<void>,
+    errorHandler?: () => void
+  ): Promise<void> => {
+    try {
+      const response = await fetch(url);
+      if (response.status === 200) {
+        void callback(
+          new File([await response.blob()], 'Preview', {
+            type,
+          })
+        );
+      } else {
         errorHandler ? errorHandler() : showSnackbarError();
       }
-    },
-    [showSnackbarError]
-  );
+    } catch {
+      errorHandler ? errorHandler() : showSnackbarError();
+    }
+  };
 
   return retrieveFile;
 };
