+++ src/components/files/hooks/index.ts
@@ -1,4 +1,3 @@
 export * from './useDownloadFile';
-export * from './useFileNameAndExtension';
 export * from './useFileNodeIcon';
 export * from './useGetFileDownloadUrl';
