+++ src/components/files/hooks/useGetFileDownloadUrl.ts
@@ -1,6 +1,5 @@
 import { useApolloClient } from '@apollo/client';
 import { useSnackbar } from 'notistack';
-import { useCallback } from 'react';
 import {
   FileDownloadUrl_Directory_Fragment,
   FileNodeDownloadUrlDocument,
@@ -10,36 +9,33 @@ import {
 export const useGetFileDownloadUrl = () => {
   const { enqueueSnackbar } = useSnackbar();
 
-  const showSnackbarError = useCallback(() => {
+  const showSnackbarError = () => {
     enqueueSnackbar('Could not download file', {
       variant: 'error',
     });
-  }, [enqueueSnackbar]);
+  };
 
   const client = useApolloClient();
 
-  const getFileDownloadUrl = useCallback(
-    async (id: string): Promise<string> => {
-      try {
-        const { data } = await client.query<FileNodeDownloadUrlQuery>({
-          query: FileNodeDownloadUrlDocument,
-          variables: { id },
-          fetchPolicy: 'network-only',
-        });
-        const isDirectory = (
-          node: FileNodeDownloadUrlQuery['fileNode'] | undefined
-        ): node is FileDownloadUrl_Directory_Fragment => {
-          return node?.type === 'Directory';
-        };
-        return !isDirectory(data?.fileNode)
-          ? data?.fileNode.downloadUrl ?? ''
-          : '';
-      } catch {
-        showSnackbarError();
-        return '';
-      }
-    },
-    [client, showSnackbarError]
-  );
+  const getFileDownloadUrl = async (id: string): Promise<string> => {
+    try {
+      const { data } = await client.query<FileNodeDownloadUrlQuery>({
+        query: FileNodeDownloadUrlDocument,
+        variables: { id },
+        fetchPolicy: 'network-only',
+      });
+      const isDirectory = (
+        node: FileNodeDownloadUrlQuery['fileNode'] | undefined
+      ): node is FileDownloadUrl_Directory_Fragment => {
+        return node?.type === 'Directory';
+      };
+      return !isDirectory(data?.fileNode)
+        ? data?.fileNode.downloadUrl ?? ''
+        : '';
+    } catch {
+      showSnackbarError();
+      return '';
+    }
+  };
   return getFileDownloadUrl;
 };
