+++ src/scenes/Projects/Files/useUploadProjectFiles.ts
@@ -1,4 +1,3 @@
-import { useCallback } from 'react';
 import { GQLOperations } from '../../../api';
 import { useCreateFileVersionMutation } from '../../../components/files/FileActions';
 import { UploadCallback, useUpload } from '../../../components/Upload';
@@ -6,24 +5,25 @@ import { UploadCallback, useUpload } from '../../../components/Upload';
 export const useHandleUploadCompleted = (parentId: string): UploadCallback => {
   const [createFileVersion] = useCreateFileVersionMutation();
 
-  const handleUploadCompleted: UploadCallback = useCallback(
-    async (uploadId, name, action) => {
-      const input = {
-        uploadId,
-        name,
-        parentId,
-      };
-      await createFileVersion({
-        variables: { input },
-        refetchQueries: [
-          action === 'version'
-            ? GQLOperations.Query.FileVersions
-            : GQLOperations.Query.ProjectDirectory,
-        ],
-      });
-    },
-    [parentId, createFileVersion]
-  );
+  const handleUploadCompleted: UploadCallback = async (
+    uploadId,
+    name,
+    action
+  ) => {
+    const input = {
+      uploadId,
+      name,
+      parentId,
+    };
+    await createFileVersion({
+      variables: { input },
+      refetchQueries: [
+        action === 'version'
+          ? GQLOperations.Query.FileVersions
+          : GQLOperations.Query.ProjectDirectory,
+      ],
+    });
+  };
   return handleUploadCompleted;
 };
 
@@ -35,20 +35,17 @@ export const useUploadProjectFiles = (
 
   const handleUploadCompleted = useHandleUploadCompleted(parentId);
 
-  const handleFilesDrop = useCallback(
-    (files: File[]) => {
-      const fileInputs = files.map((file) => ({
-        file,
-        fileName: file.name,
-        callback: (
-          uploadId: Parameters<UploadCallback>[0],
-          name: Parameters<UploadCallback>[1]
-        ) => handleUploadCompleted(uploadId, name, action),
-      }));
-      addFilesToUploadQueue(fileInputs);
-    },
-    [action, addFilesToUploadQueue, handleUploadCompleted]
-  );
+  const handleFilesDrop = (files: File[]) => {
+    const fileInputs = files.map((file) => ({
+      file,
+      fileName: file.name,
+      callback: (
+        uploadId: Parameters<UploadCallback>[0],
+        name: Parameters<UploadCallback>[1]
+      ) => handleUploadCompleted(uploadId, name, action),
+    }));
+    addFilesToUploadQueue(fileInputs);
+  };
 
   return handleFilesDrop;
 };
